Added toast message to long-press to kill

+++ policy/src/com/android/internal/policy/impl/PhoneWindowManager.java
@@ -134,6 +134,7 @@ import android.view.KeyCharacterMap.FallbackAction;
 import android.view.accessibility.AccessibilityEvent;
 import android.view.animation.Animation;
 import android.view.animation.AnimationUtils;
+import android.widget.Toast;
 import android.media.IAudioService;
 import android.media.AudioManager;
 
@@ -723,6 +724,7 @@ public class PhoneWindowManager implements WindowManagerPolicy {
     Runnable mBackLongPress = new Runnable() {
         public void run() {
             try {
+                performHapticFeedbackLw(null, HapticFeedbackConstants.LONG_PRESS, false);
                 IActivityManager mgr = ActivityManagerNative.getDefault();
                 List<RunningAppProcessInfo> apps = mgr.getRunningAppProcesses();
                 for (RunningAppProcessInfo appInfo : apps) {
@@ -731,8 +733,13 @@ public class PhoneWindowManager implements WindowManagerPolicy {
                     // root, phone, etc.)
                     if (uid >= Process.FIRST_APPLICATION_UID && uid <= Process.LAST_APPLICATION_UID
                             && appInfo.importance == RunningAppProcessInfo.IMPORTANCE_FOREGROUND) {
+                        Toast.makeText(mContext, R.string.app_killed_message, Toast.LENGTH_SHORT).show();
                         // Kill the entire pid
-                        Process.killProcess(appInfo.pid);
+                        if (appInfo.pkgList!=null && (apps.size() > 0)){
+                             mgr.forceStopPackage(appInfo.pkgList[0]);
+                             }else{
+                             Process.killProcess(appInfo.pid);
+                             }
                         break;
                     }
                 }
